ok libft

+++ ft_putunbr_base.c
@@ -6,15 +6,15 @@
 /*   By: mnieto-m <mnieto-m@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/19 20:07:44 by mnieto-m          #+#    #+#             */
-/*   Updated: 2024/06/21 13:36:19 by mnieto-m         ###   ########.fr       */
+/*   Updated: 2024/06/21 20:42:58 by mnieto-m         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	ft_putunbr_base(unsigned int n, char *base, int *count)
+void	ft_putunbr_base(size_t n, char *base, int *count)
 {
-	unsigned int len;
+	size_t len;
 
 	len = ft_strlen(base);
 		if (n > (len - 1))
